Final Recipe App/requirements.txt
mysql-connector-python==8.4.0
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
pyodbc==5.1.0
pytube==15.0.0
requests==2.32.2
ttkthemes==3.2.2
urllib3==2.2.1
bcrypt==4.2.0